+++ tests/helpers/OperatorExample.php
@@ -5,7 +5,7 @@
  * @homepage: https://github.com/andy87/Handler
  * @license CC BY-SA 4.0 http://creativecommons.org/licenses/by-sa/4.0/
  * @date 2024-05-27
- * @version 1.1.0
+ * @version 1.2.0
  */
 
 declare(strict_types=1);
@@ -13,14 +13,14 @@ declare(strict_types=1);
 namespace andy87\knock_knock\tests\helpers;
 
 use Exception;
-use andy87\knock_knock\core\Handler;
+use andy87\knock_knock\core\Operator;
 
 /**
  *  Class HandlerExample
  *
  * @package tests\examples
  */
-class HandlerExample extends Handler
+class OperatorExample extends Operator
 {
     public const INIT_INIT = 'init_Stay';
     public const INIT_DONE = 'init_Done';